requirements.txt
https://github.com/dsdanielpark/Bard-API.git
python-dotenv~=1.0.0
disnake~=2.8.1
googletrans==3.1.0a0
httpx[http2]==0.13.3
motor~=3.1.2
poe-api